+++ EveUtils.Migrations.Client.Sqlite/Migrations/20260916045526_AddActivitySummaryIskByCharacter.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace EveUtils.Migrations.Client.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddActivitySummaryIskByCharacter : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "IskContributionsByCharacter",
+                table: "ActivitySummary",
+                type: "TEXT",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IskContributionsByCharacter",
+                table: "ActivitySummary");
+        }
+    }
+}